+++ testWorkspace/glob(chars)/app.ts
@@ -1,0 +1,8 @@
+console.log("Line 1");
+console.log("Line 2");
+console.log("Line 3");
+console.log("Line 4");
+console.log("Line 5");
+console.log("Line 6");
+console.log("Line 7");
+console.log("Line 8");